# Conflicts:
#	src/scripts/everyBoot/listen-for-code.sh

+++ src/install.sh
@@ -1,18 +1,35 @@
 #!/bin/bash
 
-# Pull in global vars
-eval "$(curl -s "https://raw.githubusercontent.com/Lunars/tesla/master/src/config.sh")"
-mydir="${0%/*}"
-[ -f "$mydir/config.sh" ] && source "$mydir/config.sh"
+# Usage:
+# install.sh <customPathToHome | "cron">
 
+if [ "$EUID" -ne 0 ]; then
+  echo "[FAIL] Script must be ran as root"
+  exit
+else
+  echo "[OK] Installing Lunars as root"
+fi
+
+# Pull in global vars
+[ "$1" != "cron" ] && eval "$(curl -s "https://raw.githubusercontent.com/Lunars/tesla/master/src/config.sh")"
+
+# Set default home if we're running a cron flag
+[ "$1" == "cron" ] && homeOfLunars="/var/lunars"
+
+# The default home provided by Github config.sh
 search="$homeOfLunars"
-[ -n "$1" ] && homeOfLunars="$1"
+
+# The overridden home provided by first argument
+[ -n "$1" ] && [ "$1" != "cron" ] && homeOfLunars="$1"
+
+# Pull in your custom vars if you already have Lunars installed
+[ -f "$homeOfLunars/config.sh" ] && source "$homeOfLunars/config.sh"
 
 echo "Installing Lunars to $homeOfLunars"
 
 # Detect if we are running chrooted by checking if the root of the init process is the same as the root of this process
 if [[ "$(stat -c %d:%i /)" != "$(stat -c %d:%i /proc/1/root/.)" ]]; then
-  echo "[FAIL] Not running chrooted"
+  echo "[FAIL] Not running when chrooted"
   exit 2
 fi
 
@@ -25,10 +42,10 @@ startScript="/sbin/start-stop-daemon --start --quiet --make-pidfile --oknodo --b
 function checkConfigScripts() {
   cd "$homeOfLunars/$1" || exit
   for f in *.sh; do
-    if grep -q "$f" "$homeOfLunars/config.sh"; then
-      echo "The script [$f] already exists in config.sh."
+    if grep -q "$1/$f" "$homeOfLunars/config.sh"; then
+      echo "[SKIP] $f already exists in config.sh"
     else
-      echo "Adding [$f] to config.sh, and leaving commented out"
+      echo "[OK] Added $f to config.sh and left disabled (must manually enable)"
       sed -i "/$2=/a #    \"\$homeOfLunars/$1/$f\"" "$homeOfLunars/config.sh"
     fi
   done
@@ -38,14 +55,20 @@ function downloadLunars() {
   echo "Downloading lunars source from Github..."
 
   mkdir -p "$homeOfLunars"
-  curl -s -L https://github.com/Lunars/tesla/tarball/master | tar --wildcards -zx -C "$homeOfLunars" "Lunars-tesla-*/src"
-  mv "$homeOfLunars"/*/src/* "$homeOfLunars"
+  curl https://codeload.github.com/Lunars/tesla/legacy.tar.gz/master | tar --wildcards -zx -C "$homeOfLunars" "Lunars-tesla-*/src"
+
+  if ! mv "$homeOfLunars"/*/src/* "$homeOfLunars"; then
+    echo "[FAIL] Lunars could not be downloaded from Github"
+    exit 1
+  fi
+
   rm -rf "$homeOfLunars"/Lunars-tesla-*
 
   echo "[OK] Lunars source downloaded"
 
   if [[ -f /tmp/config.sh ]]; then
     # Restore config from tmp
+    rm -rf "$homeOfLunars/overwrite-files"
     mv /tmp/overwrite-files "$homeOfLunars"
     mv /tmp/config.sh "$homeOfLunars"
     mv /tmp/tesla.ovpn "$homeOfLunars"
@@ -57,20 +80,26 @@ function downloadLunars() {
   fi
 }
 
-if [[ -f "$onRebootFile" ]]; then
-  # Save config to tmp
-  mv "$homeOfLunars"/overwrite-files /tmp
-  mv "$homeOfLunars"/config.sh /tmp
-  mv "$homeOfLunars"/tesla.ovpn /tmp
-  echo "[OK] Lunars config.sh, tesla.ovpn, and overwrite-files saved"
+# When calling with "cron" argument, don't redownload. Just do CRON stuff
+if [ -z "$1" ] || [ -n "$1" ] && [ "$1" != "cron" ]; then
+  if [[ -f "$onRebootFile" ]]; then
+    # Save config to tmp
+    rm -rf /tmp/overwrite-files /tmp/config.sh /tmp/tesla.ovpn
+    mv "$homeOfLunars"/overwrite-files /tmp
+    mv "$homeOfLunars"/config.sh /tmp
+    mv "$homeOfLunars"/tesla.ovpn /tmp
+    echo "[OK] Lunars config.sh, tesla.ovpn, and overwrite-files saved"
 
-  rm -rf "$homeOfLunars"
-  downloadLunars
+    rm -rf "$homeOfLunars"
+    downloadLunars
+  else
+    downloadLunars
+
+    # In case a custom install path was given, replace it in config.sh too
+    [ -n "$search" ] && sed -i "s~$search~$homeOfLunars~g" "$homeOfLunars/config.sh"
+  fi
 else
-  downloadLunars
-
-  # In case a custom install path was given, replace it in config.sh too
-  [ -n "$search" ] && sed -i "s~$search~$homeOfLunars~g" "$homeOfLunars/config.sh"
+  echo "[SKIP] Lunars download due to cron flag"
 fi
 
 # Installing crontab
@@ -89,11 +118,17 @@ fi
 
 # Check if already running
 rebootProcess=$(pgrep -f "$rebootScript")
-if [ -n "$rebootProcess" ]; then
-  echo "[SKIP] Lunars $rebootScript is already running"
-else
+if [ "$1" != "cron" ] && [ -n "$rebootProcess" ]; then
+  echo "[SKIP] Lunars $rebootScript is already running, restarting process"
+  kill "$rebootProcess"
   $startScript
-  echo "[OK] Lunars $rebootScript backgrounded"
 fi
+
+if [ -z "$rebootProcess" ]; then
+  echo "[OK] Starting $rebootScript as it wasn't found to be running"
+  $startScript
+fi
+
+echo "[OK] Lunars $rebootScript backgrounded"
 
 echo "[DONE] Lunars is now installed, have fun!"
